+++ providers/alternatives.rb
@@ -20,25 +20,25 @@ include Chef::Mixin::ShellOut
 action :set do
   if new_resource.bin_cmds
     # I couldn't find a way to cleanly avoid repeating this variable declaration in both :set and :unset
-    alternatives_cmd = node['platform_family'] == 'rhel' ? 'alternatives' : 'update-alternatives'
+    alternatives_cmd = node['platform_family'] == 'rhel' ? 'alternatives' : 'update-alternatives' # rubocop: disable UselessAssignment
     new_resource.bin_cmds.each do |cmd|
 
-      bin_path = "/usr/bin/#{cmd}"
-      alt_path = "#{new_resource.java_location}/bin/#{cmd}"
-      priority = new_resource.priority
+      bin_path = '/usr/bin/#{cmd}' # rubocop: disable UselessAssignment
+      alt_path = '#{new_resource.java_location}/bin/#{cmd}'
+      priority = new_resource.priority # rubocop: disable UselessAssignment
 
-      if !::File.exist?(alt_path)
-        Chef::Log.debug "Skipping setting alternative for #{cmd}. Command #{alt_path} does not exist."
+      unless ::File.exist?(alt_path)
+        Chef::Log.debug 'Skipping setting alternative for #{cmd}. Command #{alt_path} does not exist.'
         next
       end
 
       # install the alternative if needed
-      alternative_exists = shell_out("#{alternatives_cmd} --display #{cmd} | grep #{alt_path}").exitstatus == 0
+      alternative_exists = shell_out('#{alternatives_cmd} --display #{cmd} | grep #{alt_path}').exitstatus == 0
       unless alternative_exists
-        description = "Add alternative for #{cmd}"
+        description = 'Add alternative for #{cmd}'
         converge_by(description) do
-          Chef::Log.debug "Adding alternative for #{cmd}"
-          install_cmd = shell_out("#{alternatives_cmd} --install #{bin_path} #{cmd} #{alt_path} #{priority}")
+          Chef::Log.debug 'Adding alternative for #{cmd}'
+          install_cmd = shell_out('#{alternatives_cmd} --install #{bin_path} #{cmd} #{alt_path} #{priority}')
           unless install_cmd.exitstatus == 0
             Chef::Application.fatal!(%Q[ set alternative failed ])
           end
@@ -48,13 +48,13 @@ action :set do
 
       # set the alternative if default
       if new_resource.default
-        alternative_is_set = shell_out("#{alternatives_cmd} --display #{cmd} | grep \"link currently points to #{alt_path}\"").exitstatus == 0
+        alternative_is_set = shell_out('#{alternatives_cmd} --display #{cmd} | grep \"link currently points to #{alt_path}\"').exitstatus == 0
         unless alternative_is_set
-          description = "Set alternative for #{cmd}"
+          description = 'Set alternative for #{cmd}'
           converge_by(description) do
-            Chef::Log.debug "Setting alternative for #{cmd}"
-            set_cmd = shell_out("#{alternatives_cmd} --set #{cmd} #{alt_path}")
-            unless set_cmd.exitstatus == 0
+            Chef::Log.debug 'Setting alternative for #{cmd}'
+            set_cmd = shell_out('#{alternatives_cmd} --set #{cmd} #{alt_path}')
+            unless set_cmd.exitstatus == 0 # rubocop: disable BlockNesting
               Chef::Application.fatal!(%Q[ set alternative failed ])
             end
           end
@@ -67,10 +67,10 @@ end
 
 action :unset do
   # I couldn't find a way to cleanly avoid repeating this variable declaration in both :set and :unset
-  alternatives_cmd = node['platform_family'] == 'rhel' ? 'alternatives' : 'update-alternatives'
+  alternatives_cmd = node['platform_family'] == 'rhel' ? 'alternatives' : 'update-alternatives' # rubocop: disable UselessAssignment
   new_resource.bin_cmds.each do |cmd|
-    alt_path = "#{new_resource.java_location}/bin/#{cmd}"
-    cmd = shell_out("#{alternatives_cmd} --remove #{cmd} #{alt_path}")
+    alt_path = '#{new_resource.java_location}/bin/#{cmd}' # rubocop: disable UselessAssignment
+    cmd = shell_out('#{alternatives_cmd} --remove #{cmd} #{alt_path}')
     if cmd.exitstatus == 0
       new_resource.updated_by_last_action(true)
     end