+++ recipes/openjdk.rb
@@ -2,10 +2,11 @@
 # Author:: Seth Chisamore (<schisamo@opscode.com>)
 # Author:: Joshua Timberman (<joshua@opscode.com>)
 #
-# Cookbook Name:: java
+# Cookbook Name:: rackspace_java
 # Recipe:: openjdk
 #
 # Copyright 2010-2013, Opscode, Inc.
+# Copyright 2014, Rackspace, US Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,47 +20,50 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-unless node.recipe?('java::default')
-  Chef::Log.warn("Using java::default instead is recommended.")
+unless node.recipe?('rackspace_java::default')
+  Chef::Log.warn('Using rackspace_java::default instead is recommended.')
 
   # Even if this recipe is included by itself, a safety check is nice...
-  [ node['java']['openjdk_packages'], node['java']['java_home'] ].each do |v|
-    if v.nil? or v.empty?
-      include_recipe "java::set_attributes_from_version"
+  [node['rackspace_java']['openjdk_packages'], node['rackspace_java']['java_home']].each do |v|
+    if v.nil? || v.empty?
+      include_recipe 'rackspace_java::set_attributes_from_version'
     end
   end
 end
 
 jdk = Opscode::OpenJDK.new(node)
 
-if platform_requires_license_acceptance?
-  file "/opt/local/.dlj_license_accepted" do
-    owner "root"
-    group "root"
-    mode "0400"
-    action :create
-    only_if { node['java']['accept_license_agreement'] }
-  end
-end
-
-node['java']['openjdk_packages'].each do |pkg|
+node['rackspace_java']['openjdk_packages'].each do |pkg|
   package pkg
 end
 
-if platform_family?('debian', 'rhel', 'fedora')
-  java_alternatives 'set-java-alternatives' do
-    java_location jdk.java_home
-    priority jdk.alternatives_priority
-    case node['java']['jdk_version']
-    when "6"
-      bin_cmds node['java']['jdk']['6']['bin_cmds']
-    when "7"
-      bin_cmds node['java']['jdk']['7']['bin_cmds']
-    end
-    action :set
+# if platform_family?('debian', 'rhel')
+#   rackspace_java_alternatives 'set-java-alternatives' do
+#     java_location jdk.java_home
+#     priority jdk.alternatives_priority
+#     case node['rackspace_java']['jdk_version']
+#     when '6'
+#       bin_cmds node['rackspace_java']['jdk']['6']['bin_cmds']
+#     when '7'
+#       bin_cmds node['rackspace_java']['jdk']['7']['bin_cmds']
+#     end
+#     action :set
+#   end
+# end
+
+rackspace_java_alternatives 'set-java-alternatives' do
+  java_location jdk.java_home
+  priority jdk.alternatives_priority
+  case node['rackspace_java']['jdk_version']
+  when '6'
+    bin_cmds node['rackspace_java']['jdk']['6']['bin_cmds']
+  when '7'
+    bin_cmds node['rackspace_java']['jdk']['7']['bin_cmds']
   end
+  action :set
+  only_if { platform_family?('debian', 'rhel') }
 end
 
 # We must include this recipe AFTER updating the alternatives or else JAVA_HOME
 # will not point to the correct java.
-include_recipe 'java::set_java_home'
+include_recipe 'rackspace_java::set_java_home'