+++ recipes/purge_packages.rb
@@ -1,5 +1,7 @@
-# Cookbook Name:: java
+# Cookbook Name:: rackspace_java
 # Recipe:: purge_packages
+#
+# Copyright 2014, Rackspace, US Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.