+++ recipes/set_attributes_from_version.rb
@@ -1,5 +1,7 @@
-# Cookbook Name:: java
+# Cookbook Name:: rackspace_java
 # Recipe:: set_attributes_from_version
+#
+# Copyright 2014, Rackspace, US Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,34 +20,23 @@
 # force_default or higher precedence.
 
 case node['platform_family']
-when "rhel", "fedora"
-  case node['java']['install_flavor']
-  when "oracle"
-    node.default['java']['java_home'] = "/usr/lib/jvm/java"
+when 'rhel'
+  case node['rackspace_java']['install_flavor']
+  when 'oracle'
+    node.default['rackspace_java']['java_home'] = '/usr/lib/jvm/java'
   else
-    node.default['java']['java_home'] = "/usr/lib/jvm/java-1.#{node['java']['jdk_version']}.0"
+    node.default['rackspace_java']['java_home'] = "/usr/lib/jvm/java-1.#{node['rackspace_java']['jdk_version']}.0"
   end
-  node.default['java']['openjdk_packages'] = ["java-1.#{node['java']['jdk_version']}.0-openjdk", "java-1.#{node['java']['jdk_version']}.0-openjdk-devel"]
-when "freebsd"
-  node.default['java']['java_home'] = "/usr/local/openjdk#{node['java']['jdk_version']}"
-  node.default['java']['openjdk_packages'] = ["openjdk#{node['java']['jdk_version']}"]
-when "arch"
-  node.default['java']['java_home'] = "/usr/lib/jvm/java-#{node['java']['jdk_version']}-openjdk"
-  node.default['java']['openjdk_packages'] = ["openjdk#{node['java']['jdk_version']}}"]
-when "debian"
-  node.default['java']['java_home'] = "/usr/lib/jvm/java-#{node['java']['jdk_version']}-#{node['java']['install_flavor']}"
+  node.default['rackspace_java']['openjdk_packages'] = ["java-1.#{node['rackspace_java']['jdk_version']}.0-openjdk", "java-1.#{node['rackspace_java']['jdk_version']}.0-openjdk-devel"] # rubocop: disable LineLength
+when 'debian'
+  node.default['rackspace_java']['java_home'] = "/usr/lib/jvm/java-#{node['rackspace_java']['jdk_version']}-#{node['rackspace_java']['install_flavor']}"
   # Newer Debian & Ubuntu adds the architecture to the path
-  if node['platform'] == 'debian' && Chef::VersionConstraint.new(">= 7.0").include?(node['platform_version']) ||
-     node['platform'] == 'ubuntu' && Chef::VersionConstraint.new(">= 12.04").include?(node['platform_version'])
-    node.default['java']['java_home'] = "#{node['java']['java_home']}-#{node['kernel']['machine'] == 'x86_64' ? 'amd64' : 'i386'}"
+  if node['platform'] == 'debian' && Chef::VersionConstraint.new('>= 7.0').include?(node['platform_version']) ||
+     node['platform'] == 'ubuntu' && Chef::VersionConstraint.new('>= 12.04').include?(node['platform_version'])
+    node.default['rackspace_java']['java_home'] = "#{node['rackspace_java']['java_home']}-#{node['kernel']['machine'] == 'x86_64' ? 'amd64' : 'i386'}"
   end
-  node.default['java']['openjdk_packages'] = ["openjdk-#{node['java']['jdk_version']}-jdk", "openjdk-#{node['java']['jdk_version']}-jre-headless"]
-when "smartos"
-  node.default['java']['java_home'] = "/opt/local/java/sun6"
-  node.default['java']['openjdk_packages'] = ["sun-jdk#{node['java']['jdk_version']}", "sun-jre#{node['java']['jdk_version']}"]
-when "windows"
-  # Do nothing otherwise we will fall through to the else and set java_home to an invalid path, causing the installer to popup a dialog
+  node.default['rackspace_java']['openjdk_packages'] = ["openjdk-#{node['rackspace_java']['jdk_version']}-jdk", "openjdk-#{node['rackspace_java']['jdk_version']}-jre-headless"]
 else
-  node.default['java']['java_home'] = "/usr/lib/jvm/default-java"
-  node.default['java']['openjdk_packages'] = ["openjdk-#{node['java']['jdk_version']}-jdk"]
+  node.default['rackspace_java']['java_home'] = '/usr/lib/jvm/default-java'
+  node.default['rackspace_java']['openjdk_packages'] = ["openjdk-#{node['rackspace_java']['jdk_version']}-jdk"]
 end