+++ resources/alternatives.rb
@@ -1,6 +1,8 @@
 #
-# Cookbook Name:: java
+# Cookbook Name:: rackspace_java
 # Provider:: alternatives
+#
+# Copyright 2014, Rackspace, US Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,10 +18,10 @@
 
 actions :set, :unset
 
-attribute :java_location, :kind_of => String, :default => nil
-attribute :bin_cmds, :kind_of => Array, :default => nil
-attribute :default, :equal_to => [true, false], :default => true
-attribute :priority, :kind_of => Integer, :default => 1061
+attribute :java_location, kind_of: String, default: nil
+attribute :bin_cmds, kind_of: Array, default: nil
+attribute :default, equal_to: [true, false], default: true
+attribute :priority, kind_of: Integer, default: 1061
 
 # we have to set default for the supports attribute
 # in initializer since it is a 'reserved' attribute name