+++ spec/libraries/helpers_spec.rb
@@ -15,15 +15,15 @@
 # limitations under the License.
 #
 
-$:.unshift(File.join(File.dirname(__FILE__), '..'))
+$:.unshift(File.join(File.dirname(__FILE__), '..')) # rubocop: disable SpecialGlobalVars
 require 'spec_helper'
 
 describe Opscode::OpenJDK do
   let(:node) do
     {
-      'java' => {
+      'rackspace_java' => {
         'java_home' => '/usr/lib/jvm/default-java',
-        'jdk_version' => 6
+        'jdk_version' => '6'
       },
       'kernel' => {
         'machine' => 'x86_64'
@@ -42,11 +42,11 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets java_home' do
-      expect(subject.java_home).to eq(node['java']['java_home'])
+      expect(subject.java_home).to eq(node['rackspace_java']['java_home'])
     end
 
     it 'sets jdk_version' do
-      expect(subject.jdk_version).to eq(node['java']['jdk_version'])
+      expect(subject.jdk_version).to eq(node['rackspace_java']['jdk_version'])
     end
   end
 
@@ -61,40 +61,12 @@ describe Opscode::OpenJDK do
     end
   end
 
-  context 'ubuntu 10.04 32 bit' do
-    before do
-      node['platform'] = 'ubuntu'
-      node['platform_version'] = '10.04'
-      node['platform_family'] = 'debian'
-      node['kernel']['machine'] = 'i386'
-    end
-
-    it 'sets the java location for JDK 6' do
-      expected_path = '/usr/lib/jvm/java-6-openjdk/jre/bin/java'
-      expect(subject.java_location).to eq(expected_path)
-    end
-  end
-
-  context 'ubuntu 10.04 64 bit' do
-    before do
-      node['platform'] = 'ubuntu'
-      node['platform_version'] = '10.04'
-      node['platform_family'] = 'debian'
-      node['kernel']['machine'] = 'x86_64'
-    end
-
-    it 'sets the java location for 64 bit and JDK 6' do
-      expected_path = '/usr/lib/jvm/java-6-openjdk/jre/bin/java'
-      expect(subject.java_location).to eq(expected_path)
-    end
-  end
-
   context 'ubuntu 12.04 32 bit' do
     before do
       node['platform'] = 'ubuntu'
       node['platform_version'] = '12.04'
       node['platform_family'] = 'debian'
-      node['kernel']['machine'] = 'i386'
+      node['kernel']['machine'] = 'i586'
     end
 
     it 'sets the java location for JDK 6' do
@@ -103,7 +75,7 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets the java location for JDK 7' do
-      node['java']['jdk_version'] = '7'
+      node['rackspace_java']['jdk_version'] = '7'
       expected_path = '/usr/lib/jvm/java-7-openjdk-i386/jre/bin/java'
       expect(subject.java_location).to eq(expected_path)
     end
@@ -123,36 +95,8 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets the java location for JDK 7' do
-      node['java']['jdk_version'] = '7'
+      node['rackspace_java']['jdk_version'] = '7'
       expected_path = '/usr/lib/jvm/java-7-openjdk-amd64/jre/bin/java'
-      expect(subject.java_location).to eq(expected_path)
-    end
-  end
-
-  context 'debian 6 32 bit' do
-    before do
-      node['platform'] = 'debian'
-      node['platform_version'] = '6.0.7'
-      node['platform_family'] = 'debian'
-      node['kernel']['machine'] = 'i386'
-    end
-
-    it 'sets the java location for JDK 6' do
-      expected_path = '/usr/lib/jvm/java-6-openjdk/jre/bin/java'
-      expect(subject.java_location).to eq(expected_path)
-    end
-  end
-
-  context 'debian 6 64 bit' do
-    before do
-      node['platform'] = 'debian'
-      node['platform_version'] = '6.0.7'
-      node['platform_family'] = 'debian'
-      node['kernel']['machine'] = 'x86_64'
-    end
-
-    it 'sets the java location for 64 bit and JDK 6' do
-      expected_path = '/usr/lib/jvm/java-6-openjdk/jre/bin/java'
       expect(subject.java_location).to eq(expected_path)
     end
   end
@@ -171,7 +115,7 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets the java location for JDK 7' do
-      node['java']['jdk_version'] = '7'
+      node['rackspace_java']['jdk_version'] = '7'
       expected_path = '/usr/lib/jvm/java-7-openjdk-i386/jre/bin/java'
       expect(subject.java_location).to eq(expected_path)
     end
@@ -191,7 +135,7 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets the java location for JDK 7' do
-      node['java']['jdk_version'] = '7'
+      node['rackspace_java']['jdk_version'] = '7'
       expected_path = '/usr/lib/jvm/java-7-openjdk-amd64/jre/bin/java'
       expect(subject.java_location).to eq(expected_path)
     end
@@ -211,7 +155,7 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets the java location for JDK 7' do
-      node['java']['jdk_version'] = '7'
+      node['rackspace_java']['jdk_version'] = '7'
       expected_path = '/usr/lib/jvm/jre-1.7.0-openjdk/bin/java'
       expect(subject.java_location).to eq(expected_path)
     end
@@ -231,7 +175,7 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets the java location for JDK 7' do
-      node['java']['jdk_version'] = '7'
+      node['rackspace_java']['jdk_version'] = '7'
       expected_path = '/usr/lib/jvm/jre-1.7.0-openjdk.x86_64/bin/java'
       expect(subject.java_location).to eq(expected_path)
     end
@@ -251,7 +195,7 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets the java location for JDK 7' do
-      node['java']['jdk_version'] = '7'
+      node['rackspace_java']['jdk_version'] = '7'
       expected_path = '/usr/lib/jvm/jre-1.7.0-openjdk/bin/java'
       expect(subject.java_location).to eq(expected_path)
     end
@@ -271,7 +215,7 @@ describe Opscode::OpenJDK do
     end
 
     it 'sets the java location for JDK 7' do
-      node['java']['jdk_version'] = '7'
+      node['rackspace_java']['jdk_version'] = '7'
       expected_path = '/usr/lib/jvm/jre-1.7.0-openjdk.x86_64/bin/java'
       expect(subject.java_location).to eq(expected_path)
     end