+++ CHANGELOG.md
@@ -2,6 +2,13 @@ Java Cookbook CHANGELOG
 =======================
 This file is used to list changes made in each version of the Java cookbook.
 
+v2.0.0
+------
+- removed support for OS's other than ubuntu, debian and centos/rhel
+- removed support for ibm version of java
+- namespace change to rackspace_java
+- lots of changes to make rubocop happy
+ 
 v1.18.0
 -------
 ### Improvement